impl: add web_poseidon test
