+++ crates/prover/src/core/backend/web/circle.rs
@@ -1,3 +1,4 @@
+use super::utils::convert_web_to_simd_column;
 use super::WebBackend;
 use crate::core::backend::simd::SimdBackend;
 use crate::core::backend::Col;
@@ -9,69 +10,6 @@ use crate::core::poly::circle::{
 };
 use crate::core::poly::twiddles::TwiddleTree;
 use crate::core::poly::BitReversedOrder;
-
-// WARNING: This works because they are literally the same object layout.
-//
-// The only difference is the backend methods.
-// When we implement all methods for WebGPU,
-// we will no longer need this to convert back/forth.
-pub fn convert_web_to_simd_column(col: Col<WebBackend, BaseField>) -> Col<SimdBackend, BaseField> {
-    assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-    assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-    unsafe { std::mem::transmute(col) }
-}
-
-impl Into<CircleEvaluation<WebBackend, BaseField, BitReversedOrder>>
-    for CircleEvaluation<SimdBackend, BaseField, BitReversedOrder>
-{
-    fn into(self) -> CircleEvaluation<WebBackend, BaseField, BitReversedOrder> {
-        assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-        assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl Into<CircleEvaluation<SimdBackend, BaseField, BitReversedOrder>>
-    for CircleEvaluation<WebBackend, BaseField, BitReversedOrder>
-{
-    fn into(self) -> CircleEvaluation<SimdBackend, BaseField, BitReversedOrder> {
-        assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-        assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl AsRef<TwiddleTree<SimdBackend>> for TwiddleTree<WebBackend> {
-    fn as_ref(&self) -> &TwiddleTree<SimdBackend> {
-        assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-        assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl AsRef<CirclePoly<SimdBackend>> for CirclePoly<WebBackend> {
-    fn as_ref(&self) -> &CirclePoly<SimdBackend> {
-        assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-        assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl Into<CirclePoly<WebBackend>> for CirclePoly<SimdBackend> {
-    fn into(self) -> CirclePoly<WebBackend> {
-        assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-        assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl Into<TwiddleTree<WebBackend>> for TwiddleTree<SimdBackend> {
-    fn into(self) -> TwiddleTree<WebBackend> {
-        assert_eq!(std::mem::size_of::<SimdBackend>(), 0);
-        assert_eq!(std::mem::size_of::<WebBackend>(), 0);
-        unsafe { std::mem::transmute(self) }
-    }
-}
 
 impl PolyOps for WebBackend {
     // The twiddles type is i32, and not BaseField. This is because the fast AVX mul implementation
