+++ crates/prover/src/core/backend/web/fri.rs
@@ -1,4 +1,5 @@
 use super::WebBackend;
+use crate::core::backend::simd::SimdBackend;
 use crate::core::fields::qm31::SecureField;
 use crate::core::fri::FriOps;
 use crate::core::poly::circle::SecureEvaluation;
@@ -6,15 +7,13 @@ use crate::core::poly::line::LineEvaluation;
 use crate::core::poly::twiddles::TwiddleTree;
 use crate::core::poly::BitReversedOrder;
 
-// TODO(andrew) Is this optimized?
-#[allow(unused_variables)]
 impl FriOps for WebBackend {
     fn fold_line(
         eval: &LineEvaluation<Self>,
         alpha: SecureField,
         twiddles: &TwiddleTree<Self>,
     ) -> LineEvaluation<Self> {
-        todo!()
+        SimdBackend::fold_line(eval.as_ref(), alpha, twiddles.as_ref()).into()
     }
 
     fn fold_circle_into_line(
@@ -23,12 +22,120 @@ impl FriOps for WebBackend {
         alpha: SecureField,
         twiddles: &TwiddleTree<Self>,
     ) {
-        todo!()
+        SimdBackend::fold_circle_into_line(dst.as_mut(), src.as_ref(), alpha, twiddles.as_ref());
     }
 
     fn decompose(
         eval: &SecureEvaluation<Self, BitReversedOrder>,
     ) -> (SecureEvaluation<Self, BitReversedOrder>, SecureField) {
-        todo!()
+        let (g, lambda) = SimdBackend::decompose(eval.as_ref());
+        (g.into(), lambda)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use itertools::Itertools;
+    use num_traits::One;
+    use rand::rngs::SmallRng;
+    use rand::{Rng, SeedableRng};
+
+    use crate::core::backend::simd::column::BaseColumn;
+    use crate::core::backend::simd::SimdBackend;
+    use crate::core::backend::{Column, CpuBackend};
+    use crate::core::fields::m31::BaseField;
+    use crate::core::fields::qm31::SecureField;
+    use crate::core::fields::secure_column::SecureColumnByCoords;
+    use crate::core::fri::FriOps;
+    use crate::core::poly::circle::{CanonicCoset, CirclePoly, PolyOps, SecureEvaluation};
+    use crate::core::poly::line::{LineDomain, LineEvaluation};
+    use crate::core::poly::BitReversedOrder;
+    use crate::qm31;
+
+    #[test]
+    fn test_fold_line() {
+        const LOG_SIZE: u32 = 7;
+        let mut rng = SmallRng::seed_from_u64(0);
+        let values = (0..1 << LOG_SIZE).map(|_| rng.gen()).collect_vec();
+        let alpha = qm31!(1, 3, 5, 7);
+        let domain = LineDomain::new(CanonicCoset::new(LOG_SIZE + 1).half_coset());
+        let cpu_fold = CpuBackend::fold_line(
+            &LineEvaluation::new(domain, values.iter().copied().collect()),
+            alpha,
+            &CpuBackend::precompute_twiddles(domain.coset()),
+        );
+
+        let avx_fold = SimdBackend::fold_line(
+            &LineEvaluation::new(domain, values.iter().copied().collect()),
+            alpha,
+            &SimdBackend::precompute_twiddles(domain.coset()),
+        );
+
+        assert_eq!(cpu_fold.values.to_vec(), avx_fold.values.to_vec());
+    }
+
+    #[test]
+    fn test_fold_circle_into_line() {
+        const LOG_SIZE: u32 = 7;
+        let values: Vec<SecureField> = (0..(1 << LOG_SIZE))
+            .map(|i| qm31!(4 * i, 4 * i + 1, 4 * i + 2, 4 * i + 3))
+            .collect();
+        let alpha = qm31!(1, 3, 5, 7);
+        let circle_domain = CanonicCoset::new(LOG_SIZE).circle_domain();
+        let line_domain = LineDomain::new(circle_domain.half_coset);
+        let mut cpu_fold = LineEvaluation::new(
+            line_domain,
+            SecureColumnByCoords::zeros(1 << (LOG_SIZE - 1)),
+        );
+        CpuBackend::fold_circle_into_line(
+            &mut cpu_fold,
+            &SecureEvaluation::new(circle_domain, values.iter().copied().collect()),
+            alpha,
+            &CpuBackend::precompute_twiddles(line_domain.coset()),
+        );
+
+        let mut simd_fold = LineEvaluation::new(
+            line_domain,
+            SecureColumnByCoords::zeros(1 << (LOG_SIZE - 1)),
+        );
+        SimdBackend::fold_circle_into_line(
+            &mut simd_fold,
+            &SecureEvaluation::new(circle_domain, values.iter().copied().collect()),
+            alpha,
+            &SimdBackend::precompute_twiddles(line_domain.coset()),
+        );
+
+        assert_eq!(cpu_fold.values.to_vec(), simd_fold.values.to_vec());
+    }
+
+    #[test]
+    fn decomposition_test() {
+        const DOMAIN_LOG_SIZE: u32 = 5;
+        const DOMAIN_LOG_HALF_SIZE: u32 = DOMAIN_LOG_SIZE - 1;
+        let s = CanonicCoset::new(DOMAIN_LOG_SIZE);
+        let domain = s.circle_domain();
+        let mut coeffs = BaseColumn::zeros(1 << DOMAIN_LOG_SIZE);
+        // Polynomial is out of FFT space.
+        coeffs.as_mut_slice()[1 << DOMAIN_LOG_HALF_SIZE] = BaseField::one();
+        let poly = CirclePoly::<SimdBackend>::new(coeffs);
+        let values = poly.evaluate(domain);
+        let avx_column = SecureColumnByCoords::<SimdBackend> {
+            columns: [
+                values.values.clone(),
+                values.values.clone(),
+                values.values.clone(),
+                values.values.clone(),
+            ],
+        };
+        let avx_eval = SecureEvaluation::new(domain, avx_column.clone());
+        let cpu_eval =
+            SecureEvaluation::<CpuBackend, BitReversedOrder>::new(domain, avx_eval.values.to_cpu());
+        let (cpu_g, cpu_lambda) = CpuBackend::decompose(&cpu_eval);
+        let (avx_g, avx_lambda) = SimdBackend::decompose(&avx_eval);
+
+        assert_eq!(avx_lambda, cpu_lambda);
+        for i in 0..1 << DOMAIN_LOG_SIZE {
+            assert_eq!(avx_g.values.at(i), cpu_g.values.at(i));
+        }
     }
 }
