+++ crates/prover/src/core/backend/web/lookups/mod.rs
@@ -1,2 +1,3 @@
 mod gkr;
 mod mle;
+mod utils;
